Add support for exhibitor
Let the zookeeper process and config be managed by an existing install
of exhibitor (this will not install or manage the exhibitor service).

If `$service_provider` is set to `exhibitor` do not create the `zoo.cfg` or
`myid` files, and do not configure a system service. Exhibitor will
create the config file and manage the zookeeper process.

https://github.com/soabase/exhibitor